+++ src/main/java/com/halo/entity/UserAddress.java
@@ -6,27 +6,27 @@ import java.util.Date;
  * @author MelloChan
  */
 public class UserAddress {
-  private Long id;
-  private Long userId;
+  private Integer id;
+  private Integer userId;
   private String userName;
   private String userPhone;
   private String userAddress;
   private Date gmtCreate;
   private Date gmtUpdated;
 
-  public Long getId() {
+  public Integer getId() {
     return id;
   }
 
-  public void setId(Long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
 
-  public Long getUserId() {
+  public Integer getUserId() {
     return userId;
   }
 
-  public void setUserId(Long userId) {
+  public void setUserId(Integer userId) {
     this.userId = userId;
   }
